move unit tests to ethereum/tests
